sesion_05/requirements.txt
Django==4.0.3
djangorestframework==3.13.1
Pillow==9.1.0
